requirements.txt
django>=1.5.1
django-floppyforms>=1.1
